+++ config/configload/load.go
@@ -154,12 +154,12 @@ func bodiesToConfig(parsedBodies []*hclsyntax.Body, srcBytes [][]byte, env strin
 
 	settingsBlock := mergeSettings(parsedBodies)
 
-	definitionsBlock, err := mergeDefinitions(parsedBodies)
+	definitionsBlock, proxies, err := mergeDefinitions(parsedBodies)
 	if err != nil {
 		return nil, err
 	}
 
-	serverBlocks, err := mergeServers(parsedBodies)
+	serverBlocks, err := mergeServers(parsedBodies, proxies)
 	if err != nil {
 		return nil, err
 	}