+++ config/configload/merge.go
@@ -18,7 +18,7 @@ const (
 	errUniqueLabels     = "All %s blocks must have unique labels."
 )
 
-func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
+func mergeServers(bodies []*hclsyntax.Body, proxies map[string]*hclsyntax.Block) (hclsyntax.Blocks, error) {
 	type (
 		namedBlocks   map[string]*hclsyntax.Block
 		apiDefinition struct {
@@ -151,6 +151,10 @@ func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
 						return nil, newMergeError(errUniqueLabels, block)
 					}
 
+					if err := addProxy(block, proxies); err != nil {
+						return nil, err
+					}
+
 					results[serverKey].endpoints[block.Labels[0]] = block
 				} else if block.Type == api {
 					var apiKey string
@@ -196,6 +200,10 @@ func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
 
 							if len(subBlock.Labels) == 0 {
 								return nil, newMergeError(errUniqueLabels, subBlock)
+							}
+
+							if err := addProxy(subBlock, proxies); err != nil {
+								return nil, err
 							}
 
 							results[serverKey].apis[apiKey].endpoints[subBlock.Labels[0]] = subBlock
@@ -401,11 +409,12 @@ func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
 	return mergedServers, nil
 }
 
-func mergeDefinitions(bodies []*hclsyntax.Body) (*hclsyntax.Block, error) {
+func mergeDefinitions(bodies []*hclsyntax.Body) (*hclsyntax.Block, map[string]*hclsyntax.Block, error) {
 	type data map[string]*hclsyntax.Block
 	type list map[string]data
 
 	definitionsBlock := make(list)
+	proxiesList := make(data)
 
 	for _, body := range bodies {
 		for _, outerBlock := range body.Blocks {
@@ -415,8 +424,6 @@ func mergeDefinitions(bodies []*hclsyntax.Body) (*hclsyntax.Block, error) {
 						definitionsBlock[innerBlock.Type] = make(data)
 					}
 
-					definitionsBlock[innerBlock.Type][innerBlock.Labels[0]] = innerBlock
-
 					// Count the "backend" blocks and "backend" attributes to
 					// forbid multiple backend definitions.
 
@@ -425,7 +432,7 @@ func mergeDefinitions(bodies []*hclsyntax.Body) (*hclsyntax.Block, error) {
 					for _, block := range innerBlock.Body.Blocks {
 						if block.Type == errorHandler {
 							if err := absInBackends(block); err != nil {
-								return nil, err
+								return nil, nil, err
 							}
 						} else if block.Type == backend {
 							backends++
@@ -437,7 +444,28 @@ func mergeDefinitions(bodies []*hclsyntax.Body) (*hclsyntax.Block, error) {
 					}
 
 					if backends > 1 {
-						return nil, newMergeError(errMultipleBackends, innerBlock)
+						return nil, nil, newMergeError(errMultipleBackends, innerBlock)
+					}
+
+					if innerBlock.Type != proxy {
+						definitionsBlock[innerBlock.Type][innerBlock.Labels[0]] = innerBlock
+					} else {
+						label := innerBlock.Labels[0]
+
+						if attr, ok := innerBlock.Body.Attributes["name"]; ok {
+							name, err := attrStringValue(attr)
+							if err != nil {
+								return nil, nil, err
+							}
+
+							innerBlock.Labels[0] = name
+
+							delete(innerBlock.Body.Attributes, "name")
+						} else {
+							innerBlock.Labels[0] = defaultNameLabel
+						}
+
+						proxiesList[label] = innerBlock
 					}
 				}
 			}
@@ -457,7 +485,7 @@ func mergeDefinitions(bodies []*hclsyntax.Body) (*hclsyntax.Block, error) {
 		Body: &hclsyntax.Body{
 			Blocks: blocks,
 		},
-	}, nil
+	}, proxiesList, nil
 }
 
 func mergeDefaults(bodies []*hclsyntax.Body) (*hclsyntax.Block, error) {
@@ -586,6 +614,41 @@ func absInBackends(block *hclsyntax.Block) error {
 	}
 
 	return nil
+}
+
+func addProxy(block *hclsyntax.Block, proxies map[string]*hclsyntax.Block) error {
+	if attr, ok := block.Body.Attributes[proxy]; ok {
+		reference, err := attrStringValue(attr)
+		if err != nil {
+			return err
+		}
+
+		if proxyBlock, ok := proxies[reference]; !ok {
+			sr := attr.Expr.StartRange()
+
+			return newDiagErr(&sr, "proxy reference is not defined")
+		} else {
+			delete(block.Body.Attributes, proxy)
+
+			block.Body.Blocks = append(block.Body.Blocks, proxyBlock)
+		}
+	}
+
+	return nil
+}
+
+func attrStringValue(attr *hclsyntax.Attribute) (string, error) {
+	v, err := eval.Value(nil, attr.Expr)
+	if err != nil {
+		return "", err
+	}
+
+	if v.Type() != cty.String {
+		sr := attr.Expr.StartRange()
+		return "", newDiagErr(&sr, fmt.Sprintf("%s must evaluate to string", attr.Name))
+	}
+
+	return v.AsString(), nil
 }
 
 // newErrorHandlerKey returns a merge key based on a possible mixed error-kind format.